requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.11.3
PyYAML==6.0.2
python-multipart==0.0.20
